backend/requirements.txt
Flask
flask-cors
gunicorn
Pillow
numpy<2.0
